Add viewbox line to the basic demo
This is in response to #92 , It's been a while and @sdiemert hasn't submitted a PR, so I thought I would.

I don't think it needs to be explained here, as the demo is trying to be super minimal. However, it might be worth explaining it elsewhere. Do you have any thoughts on where?